Add support for passing settings to the server
Settings can be passed to the server through
`workspace/didChangeConfiguration` notifications:

https://github.com/eclipse/eclipse.jdt.ls/issues/1131
https://github.com/eclipse/eclipse.jdt.ls/issues/1581

It is not clear if all or only some settings can be configured this way
but I have tried setting the `java.format.settings.url` and
`java.format.settings.profile` with success. It is possible that all
configuration properties in
https://github.com/redhat-developer/vscode-java/blob/master/package.json
are supported.

The code is adapted from nvim-lspconfig and sends a notification once on
initialization of the server. But I have very little experience with
both LSP and Lua so this may have to be tweaked for a more general
solution.