Add workspace/configuration handler that recognizes insertSpaces/tabSize
If using a eclipse.jdt.ls version that includes
https://github.com/eclipse/eclipse.jdt.ls/pull/1657 this will finally
make code-actions respect the clients shiftwidth/expandtab options and
no longer include tabs when the user uses spaces.

The alternative to overriding the handler would have been to document
that users should include these options in their `config.settings`, but
given that there is no good reason that I can think of where you don't
want to respond with your local editor settings it seems justified to
include this out of the box instead of requiring every user to add this.