Add way to generate main->test code mapping via coverage & show affected
Allows to generate a coverage mapping from main code to test code by
running each test suite one after another, each time processing the
generated coverage data.

After that the coverage data can be combined with a git diff to see
which tests are most likely to be affected.

Not bullet proof, but can in many cases safe the time to run the full
test suite, reducing the feedback loop.